Major Updates

+++ app/lib/fileupload.php
@@ -89,7 +89,7 @@ class FileUpload
 
     public function remove($file)
     {
-        if(file_exists(IMAGES_UPLOAD_STORAGE.DS.$file) && is_writable(IMAGES_UPLOAD_STORAGE)) {
+        if($file !== '' && file_exists(IMAGES_UPLOAD_STORAGE.DS.$file) && is_writable(IMAGES_UPLOAD_STORAGE)) {
             unlink(IMAGES_UPLOAD_STORAGE.DS.$file);
         }
     }
