+++ app/views/clients/default.view.php
@@ -24,9 +24,15 @@
                 <td class="controls_td">
                     <a href="javascript:;" class="open_controls"><i class="fa fa-caret-square-o-left"></i></a>
                     <div class="controls_container">
-                        <a href="/clients/view/<?= $client->id ?>"><i class="fa fa-eye"></i> <?= $text_table_control_view ?></a>
-                        <a href="/clients/edit/<?= $client->id ?>"><i class="fa fa-edit"></i> <?= $text_table_control_edit ?></a>
-                        <a href="/clients/delete/<?= $client->id ?>" onclick="if(!confirm('<?= $text_table_control_delete_confirm ?>')) return false;"><i class="fa fa-trash"></i> <?= $text_table_control_delete ?></a>
+                        <?php if(array_key_exists('/clients/view', $__privilegesKeys)): ?>
+                            <a href="/clients/view/<?= $client->id ?>"><i class="fa fa-eye"></i> <?= $text_table_control_view ?></a>
+                        <?php endif; ?>
+                        <?php if(array_key_exists('/clients/edit', $__privilegesKeys)): ?>
+                            <a href="/clients/edit/<?= $client->id ?>"><i class="fa fa-edit"></i> <?= $text_table_control_edit ?></a>
+                        <?php endif; ?>
+                        <?php if(array_key_exists('/clients/delete', $__privilegesKeys)): ?>
+                            <a href="/clients/delete/<?= $client->id ?>" onclick="if(!confirm('<?= $text_table_control_delete_confirm ?>')) return false;"><i class="fa fa-trash"></i> <?= $text_table_control_delete ?></a>
+                        <?php endif; ?>
                     </div>
                 </td>
             </tr>
